requirements.txt
# Core dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0

# AI/LLM packages
openai>=1.12.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99

# LangChain packages
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

# Document processing
pypdf2>=3.0.0
python-docx>=1.0.0

# Data validation
pydantic>=2.0.0
